requirements.txt
Flask==0.10.1
Flask-Cors==2.0.1
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
SQLAlchemy==1.0.2
num2words==0.5.2
psycopg2==2.6
python-dateutil==2.4.2
pytz==2015.4
time2words==0.1.2
